+++ src/tensors/cpu/prod.cpp
@@ -15,7 +15,7 @@
 #endif
 #endif
 
-#include "sharp/int_gemm.h"
+#include "integer_common.h"
 
 namespace marian {
 
@@ -164,7 +164,7 @@ void ProdWithBias(marian::Tensor C,
                   float beta,
                   float scalar) {
   cpu::Prod(C, A, B, transA, transB, beta, scalar);
-  cpu::int16::AddBias(C, bias);
+  cpu::integer::AddBias(C, bias);
 }
 
 void CSRProd(marian::Tensor C,
